Changed the output format.
I have changed the output format of pyknogram(). The feature vectors are
of size nChannels. Each freq. bin contains a count of the number of estimated
frequencies that fall within a certain distance (BW/10) of the center frequency
of that bin divided by the bandwidth. This results in a very "clean" representation
of pyknograms that fills well with the terminology.